FE: Rename SideNavContextMenu -> SideNav
