+++ frontend/src/lib/figma/overlays/context-menu/SideNav.svelte
