+++ stories/src/stories/figma/overlays/context-menu/SideNav
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 /*
- *  Copyright (C) 2023 JWP Consulting GK
+ *  Copyright (C) 2023-2024 JWP Consulting GK
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as published
@@ -16,17 +16,17 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import SideNavContextMenu from "$lib/figma/overlays/context-menu/SideNavContextMenu.svelte";
+import SideNav from "$lib/figma/overlays/context-menu/SideNav.svelte";
 import { workspace } from "$lib-stories/storybook";
 import type { Meta, StoryObj } from "@storybook/svelte";
 
-const meta: Meta<SideNavContextMenu> = {
-    component: SideNavContextMenu,
+const meta: Meta<SideNav> = {
+    component: SideNav,
     argTypes: {},
     args: { workspace },
 };
 export default meta;
 
-type Story = StoryObj<SideNavContextMenu>;
+type Story = StoryObj<SideNav>;
 
 export const Default: Story = {};
